+++ UI12/src/app/app-routing.module.ts
@@ -1,0 +1,24 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { OrbitComponent } from './nonportal/orbit/orbit.component';
+import { Uni1993Component } from './nonportal/uni1993/uni1993.component';
+import { UnitopiaComponent } from './nonportal/unitopia/unitopia.component';
+
+const routes: Routes = [
+  { path: 'orbit',   /* canActivate: [NonPortalGuard], */component: OrbitComponent },
+  { path: 'uni1993',   /* canActivate: [NonPortalGuard], */component: Uni1993Component },
+  /*{ path: 'portal',  canActivate: [PortalGuard],component: UnitopiaComponent },*/
+  { path: '',       /* canActivate: [NonPortalGuard], */component: UnitopiaComponent,  pathMatch: 'full' },
+  { path: '**', redirectTo: '/'}
+];
+
+@NgModule({
+  imports: [
+    RouterModule.forRoot(
+      routes,
+      { enableTracing: false } // <-- debugging purposes only
+    )
+  ],
+  exports: [RouterModule]
+})
+export class AppRoutingModule { }