+++ UI12/src/app/app.module.ts
@@ -1,0 +1,43 @@
+import { BrowserModule } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
+import { CookieService } from 'ngx-cookie-service';
+
+import { AppRoutingModule } from './app-routing.module';
+import { AppComponent } from './app.component';
+import { MudModule } from './mud/mud.module';
+import { WINDOW_PROVIDERS } from './shared/WINDOW_PROVIDERS';
+import { NonportalModule } from './nonportal/nonportal.module';
+import { PrimeModule } from './prime.module';
+import { HttpClientModule } from '@angular/common/http';
+import { ModelessModule } from './modeless/modeless.module';
+
+import { AceModule } from 'ngx-ace-wrapper';
+import { ACE_CONFIG } from 'ngx-ace-wrapper';
+import { AceConfigInterface } from 'ngx-ace-wrapper';
+
+const DEFAULT_ACE_CONFIG: AceConfigInterface = {
+};
+
+@NgModule({
+  declarations: [
+    AppComponent
+  ],
+  imports: [
+    BrowserModule,HttpClientModule,
+    MudModule,NonportalModule,
+    PrimeModule,ModelessModule,
+    AppRoutingModule,
+    AceModule
+  ],
+  providers: [
+    WINDOW_PROVIDERS,
+    CookieService,
+    
+    {
+      provide: ACE_CONFIG,
+      useValue: DEFAULT_ACE_CONFIG
+    }
+  ],
+  bootstrap: [AppComponent]
+})
+export class AppModule { }