+++ UI12/src/app/menu/menu.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { MainMenuComponent } from './main-menu/main-menu.component';
+
+
+
+@NgModule({
+  declarations: [MainMenuComponent],
+  imports: [
+    CommonModule
+  ]
+})
+export class MenuModule { }