+++ UI12/src/app/mud/webmud-config.ts
@@ -1,0 +1,9 @@
+export class WebmudConfig {
+    mudname : string = '';
+    autoConnect : boolean = false;
+    autoLogin : boolean = false;
+    autoUser : string = '';
+    autoToken : string = '';
+    autoPw? :string ='';
+    localEcho : boolean = false;
+}