+++ UI12/src/locale/i18nStrings.ts
@@ -1,0 +1,6 @@
+export const i18nStrings = {
+  "de": {
+  },
+  "en": {
+  }
+}