+++ UI9/src/app/mud/mud-signals.spec.ts
@@ -1,0 +1,7 @@
+import { MudSignals } from './mud-signals';
+
+describe('MudSignals', () => {
+  it('should create an instance', () => {
+    expect(new MudSignals()).toBeTruthy();
+  });
+});