+++ UI9/src/app/mud/mud-signals.ts
@@ -1,0 +1,12 @@
+import { FileEntries } from './file-entries';
+import { FileInfo } from './file-info';
+
+export class MudSignals {
+    signal : string;
+    id : string;
+    wizard? : number;
+    playSoundFile?: string;
+    filepath?: string;
+    fileinfo? : FileInfo;
+    entries?: FileEntries[];
+}