+++ VERSION_HISTORY.md
@@ -30,3 +30,4 @@
 30. Version 0.0.44: ngx-logger inclusive logger backend, all console.logs rerouted through the new logger.
 31. Version 0.0.45: editor-save rewritten and debugged => working now.files module now mandatory. 
 32. Version 0.0.46: sqlite, nebuglog backend and frontend, start position of nonmodal windows changed, lots of bugfixes
+33. Version 0.0.47: start UI9: rewrite UI8 to primeng elements.